p-roj1/pollution_indicator/requirements.txt
flask
flask-sqlalchemy
requests
python-dotenv
flask-cors
